// frontend/src/types/html-service.types/nave-elm.type.d.ts
export type NaveElmType = {
    startNavElement: HTMLElement | null;
    plNavElement: HTMLElement | null;
    ctgAccordionElement: HTMLElement | null;
    ctgNavElement: HTMLElement | null;
    ctgIncomeNavElement: HTMLElement | null;
    ctgCostNavElement: HTMLElement | null;
    incomeNavBottom: HTMLElement | null;
};
//# sourceMappingURL=nave-elm.type.d.ts.map